requirements.txt
# Smart Microgrid Manager Pro - Deployment Requirements

# Core application dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0

# Optimization (PuLP linear programming)
pulp>=2.7.0

# Machine Learning Forecasting
scikit-learn>=1.3.0
joblib>=1.3.0

# HTTP requests (for weather API)
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0

# Report Generation
openpyxl>=3.1.0
reportlab>=4.0.0

# Session management
streamlit-authenticator>=0.2.0

# Caching and performance
diskcache>=5.6.0

# Data validation
pydantic>=2.0.0

# HTTP client
httpx>=0.25.0

# Snowflake Integration
snowflake-connector-python>=3.0.0
snowflake-snowpark-python>=1.0.0

# Additional visualizations
matplotlib>=3.7.0
seaborn>=0.12.0


